Proposed workaround for Py2.5 compatibility
This is the way to get around SyntaxError exceptions at collection
time when py.test sees this file and py2.5 is the interpreter.

The RootController that is decorated needs to live somewhere else and
imported. Not sure how to make this nested class accessing attributes
in the parent method available on demand though. It needs refactoring.